move spring repo to correct package
implements AM-941
